Fix [ Bug #204246 ] dlclose can crash or stuck system
When calling dlclose() then system can crash or freeze.
This is due to a bug in uclibc.
htab_delete() into ldso/ldso/fdpic/dl-inlines.h use size_t for i which is typedef to unsigned int.
We exit for loop on negative value of i which can never occur since i is an unsigned int !!!!!
This lead to random free of various pointers that kill system